+++ include/slick/net/http.h
@@ -21,7 +21,31 @@
 
 #pragma once
 
-#include "pch.h"
+#include <atomic>
+#include <cstdlib>
+#include <functional>
+#include <unordered_map>
+#include <mutex>
+#include <csignal>
+#include <thread>
+
+#define WIN32_LEAN_AND_MEAN
+
+#include <boost/beast/core.hpp>
+#include <boost/beast/websocket.hpp>
+#include <boost/beast/websocket/ssl.hpp>
+#include <boost/beast/http.hpp>
+#include <boost/asio/ssl.hpp>
+#include <boost/asio/strand.hpp>
+#include <boost/asio/awaitable.hpp>
+#include <boost/asio/co_spawn.hpp>
+#include <boost/asio/signal_set.hpp>
+#include <boost/asio/ip/tcp.hpp>
+#include <boost/asio/ssl/stream.hpp>
+#include <boost/asio/connect.hpp>
+#include <boost/asio/post.hpp>
+#include <boost/asio/as_tuple.hpp>
+#include <slick/slick_queue.h>
 
 namespace beast = boost::beast;         // from <boost/beast.hpp>
 namespace http = beast::http;           // from <boost/beast/http.hpp>
@@ -49,7 +73,7 @@ using tcp = boost::asio::ip::tcp;       // from <boost/asio/ip/tcp.hpp>
 #define LOG_TRACE(...) do {} while(0)
 #endif
 
-namespace slick_net {
+namespace slick::net {
 
 class Http
 {   
@@ -1184,4 +1208,4 @@ struct HttpStreamTerminater
 
 inline static HttpStreamTerminater s_http_stream_terminater;
 
-}   // namespace slick_net
+}   // namespace slick::net